Move `Edges` to `shape::Edges`
